partitioning code by generation, fixing old code
