+++ minimdo/graph/graphview.py
@@ -6,7 +6,7 @@ import numpy as np
 import sympy.printing as printing
 from anytree import RenderTree, PreOrderIter
 from trash.notationtree import all_vars_from_incidence
-from modeling.compute import INTER, END
+from modeling.gen2.nesting import INTER, END
 
 def getallvars(eqs, sympy=True):
     vrs = set()
