+++ minimdo/solver/runpipeline.py
@@ -1,5 +1,5 @@
 from graph.graphutils import flat_graph_formulation, solver_children, root_solver, Node, SOLVER, COMP, namefromsympy
-from modeling.execution import generate_components_and_residuals
+from modeling.gen2.transformations import generate_components_and_residuals
 from graph.graphutils import namefromid
 from graph.operators import sort_scc, reorder_merge_solve
 from graph.workflow import get_f, order_from_tree, default_solver_options, mdao_workflow, implicit_comp_name, OPT, SOLVE
