+++ minimdo/utils/executeformulations.py
@@ -1,6 +1,6 @@
 import numpy as np
 import openmdao.api as om
-from modeling.compute import buildidpvars
+from problemgen.gen1.openmdao import buildidpvars
 from trash.notationcompute import explicit, solver
 
 # FUNCTIONS TO MAKE THE ARCHITECTURE EXECUTABLE AND RUN IT
